api_yamdb/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2
django-filter==23.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
gunicorn==20.0.4
flake8==6.0.0
flake8-isort==6.0.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.1
pep8-naming==0.13.3
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.26.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.16
